+++ bootlocal.sh.d/01_ash.sh
@@ -9,7 +9,7 @@
 # -------------------------------------------------------------------
 
 # Backup original .ashrc and restores it when needed
-echo "== Adding complements to $HOME_DOCKER_USER/.ashrc !"
+echo "[INFO] Adding complements to $HOME_DOCKER_USER/.ashrc !"
 HOME_DOCKER_USER_ASRC="$HOME_DOCKER_USER/.ashrc"
 if [ -f $HOME_DOCKER_USER_ASRC.original ] ; then
     cp -f $HOME_DOCKER_USER_ASRC.original $HOME_DOCKER_USER_ASRC
@@ -19,4 +19,4 @@ fi
 
 # Customize .ashrc
 cat $BOOT2DOCKER_EXTENSION_DIR/ash/.ashrc >> $HOME_DOCKER_USER_ASRC
-echo "== Complements added to $HOME_DOCKER_USER/.ashrc !"
+echo "[INFO] Complements added to $HOME_DOCKER_USER/.ashrc !"