+++ bootlocal.sh.d/02_syncd.sh
@@ -3,5 +3,5 @@
 # Syncd daemon lunching
 # ===================================================================
 
-echo "== Starting syncd daemon !"
+echo "[INFO] Starting syncd daemon !"
 $BOOT2DOCKER_EXTENSION_DIR/daemons/syncd/init.d/syncd.sh start